Refactor Journey.tsx
